+++ modules/reference_store/variables.tf
@@ -1,0 +1,4 @@
+variable "bucket_name" {
+  description = "Globally-unique name for the reference-data bucket."
+  type        = string
+}